Update MemoryGame.java
11/7: 
Added a bit of description, started defining main by importing a Scanner and creating an instance of it since we'll need to read the user's inputs, and adding a print of greeting and the rules. Nothing too complicated just yet.
Some things to consider; 1). How many rounds should this go for? I think we should keep it fairly short, at least right now. We could consider testing it at 10 rounds for now; it's still fairly short but allows the user at least three attempts. And if we have enough time we can add more rounds later, but for testing purposes I'd suggest keeping them shorter.
2).  Should we only allow yes and no answers? Or 'seen' and 'haven't seen' or something along those lines? How do we handle unsupported inputs to avoid breaking the game?

These were a couple of thoughts I had while writing this bit. I'll be working on this on and off this weekend, so feel free to change/add to the things I've done, or send me a message if you want to discuss any of these things.  Let me know your thoughts.

Alice