Add files via upload
I added comments to MemoryGame. I also made it so that the user couldn't keep losing lives after the game was over. I couldn't figure out how to do it for the points though (every time I tried it broke the game), but at least we can't have -5 lives now. Also changed the rules in main to make them accurate to the code. Feel free to change anything, especially if I commented something wrong.
Still not sure how to do the words though. Will still continue to work on this, but this is what I have so far.